+++ mycodo/controllers/base_conditional.py
@@ -38,10 +38,11 @@ class AbstractConditional:
             message = self.message
         self.message = control.trigger_all_actions(self.function_id, message=message)
 
-    def run_action(self, action_id, message=None):
+    def run_action(self, action_id, value=None, message=None):
         if message is None:
             message = self.message
-        self.message = control.trigger_action(action_id, message=message, single_action=True)
+        self.message = control.trigger_action(
+            action_id, value=value, message=message, single_action=True)
 
     @staticmethod
     def condition(condition_id):